+++ src/main/java/CPU/FooRidge/dto/user/AddUserRequest.java
@@ -1,10 +1,9 @@
-package CPU.FooRidge.dto;
+package CPU.FooRidge.dto.user;
 
 import CPU.FooRidge.domain.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @AllArgsConstructor
 @NoArgsConstructor
