+++ src/main/java/CPU/FooRidge/dto/user/LoginUserRequest.java
@@ -1,4 +1,4 @@
-package CPU.FooRidge.dto;
+package CPU.FooRidge.dto.user;
 
 import CPU.FooRidge.domain.User;
 import lombok.AllArgsConstructor;
