+++ src/main/java/CPU/FooRidge/service/UserService.java
@@ -1,23 +1,14 @@
 package CPU.FooRidge.service;
 
 import CPU.FooRidge.domain.User;
-import CPU.FooRidge.dto.AddUserRequest;
-import CPU.FooRidge.dto.LoginUserRequest;
+import CPU.FooRidge.dto.user.AddUserRequest;
+import CPU.FooRidge.dto.user.LoginUserRequest;
 import CPU.FooRidge.repository.UserRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
 @Service
 public class UserService {
